+++ api/.env
@@ -1,3 +1,3 @@
-SECRET_KEY=
-MONGO_URI=
+SECRET_KEY='randomString'
+MONGO_URI='mongodb+srv://adrian:2z3zVj3M3w6kFJrL@cluster0.tyrfy.mongodb.net/master?retryWrites=true&w=majority'
 PORT=5000
