+++ mobile/src/app/components/protests/protests.component.html
@@ -3,84 +3,11 @@
     <div>
       <ion-row>
         <ion-col size="12">
-          <h2>Create Protest</h2>
-          <form [formGroup]="form" (ngSubmit)="onSubmit()">
-            <ion-item class="app-input-body-1 ion-no-padding">
-              <ion-label position="floating">Title</ion-label>
-              <ion-input
-                required
-                formControlName="title"
-                name="title"
-              ></ion-input>
-            </ion-item>
-            <ion-item class="app-input-body-1 ion-no-padding">
-              <ion-label position="floating">Description</ion-label>
-              <ion-input
-                required
-                formControlName="description"
-                name="description"
-              ></ion-input>
-            </ion-item>
-            <ion-item class="app-input-body-1 ion-no-padding">
-              <ion-label position="floating">Start Date</ion-label>
-              <ion-datetime
-                required
-                displayFormat="D MMM YYYY H:mm"
-                formControlName="startDate"
-                name="startDate"
-                min="{{ minDate }}"
-                max="2030"
-              ></ion-datetime>
-            </ion-item>
-            <ion-col>
-              <div
-                *ngIf="form.hasError('required', 'title') && title.touched"
-                class="alert alert-danger"
-              >
-                <span class="alert-text">- A Title is required.</span>
-              </div>
-              <div
-                *ngIf="
-                  form.hasError('required', 'description') &&
-                  description.touched
-                "
-                class="alert alert-danger"
-              >
-                <span class="alert-text">- A description is required.</span>
-              </div>
-              <div
-                *ngIf="
-                  form.hasError('required', 'startDate') && startDate.touched
-                "
-                class="alert alert-danger"
-              >
-                <span class="alert-text">- A start date is required. </span>
-              </div>
-            </ion-col>
-            <ion-row class="row-items-end">
-              <ion-button
-                color="app-light-green"
-                class="submit-button"
-                expand="block"
-                type="submit"
-                [disabled]="!form.valid"
-              >
-                Submit
-              </ion-button>
-              <ion-button
-                color="app-light-green"
-                class="submit-button"
-                expand="block"
-                type="reset"
-              >
-                Clear
-              </ion-button>
-            </ion-row>
-          </form>
+          <app-create-protest></app-create-protest>
         </ion-col>
         <ion-col size="12">
           <h2>Join Protest</h2>
-          <form [formGroup]="form" (ngSubmit)="onSubmit(form)">
+          <form (ngSubmit)="onSubmit(form)">
             <p class="alert-text">//Commented out due to errors 🙀☕</p>
             <!-- <ion-item class="app-input-body-1 ion-no-padding">
               <ion-label position="floating">Protest Code</ion-label>
