+++ mobile/src/app/components/protests/protests.component.scss
@@ -1,6 +1,3 @@
-.container {
-  margin: 2rem 1rem;
-}
 .protests {
   overflow-y: scroll;
 }
@@ -13,10 +10,7 @@
     background: gray;
   }
 }
-.alert-text {
-  font-size: 13px;
-  color: rgba(78, 78, 78, 0.644);
-}
+
 hr {
   margin-top: 1.5em;
   margin-bottom: 1.5em;
